core: Add circuit module
